divide into components
