+++ src/App.css
@@ -1,178 +1,8 @@
 .App {
-  text-align: center;
+    text-align: center;
 }
 
 .App-header {
-  background-color: #f2f2f2;
-  padding: 0;
-}
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  position: fixed;
-  font-size: 16px;
-  letter-spacing: 0;
-  line-height: 50px;
-  top: 0;
-  right: 0;
-  left: 0;
-  padding: 5px 100px;
-  z-index: 1;
-  background-color: transparent;
-  border: none;
-  outline: none;
-}
-
-.navbar.scrolled {
-  background-image: linear-gradient(to bottom, rgba(242, 242, 242, 0.9), rgba(0, 0, 0, 0));
-}
-
-.logo { /* hackcon */
-  font-size: 20px;
-  letter-spacing: 7px;
-  line-height: 17px;
-  font-weight: bold;
-}
-
-.nav-links {
-  cursor: pointer;
-  list-style: none;
-  display: flex;
-}
-
-.nav-links li {
-  margin-left: 35px;
-}
-
-.nav-links li a {
-  text-decoration: none;
-  color: #000000; /* Set the desired color for the links */
-}
-
-.nav-links li a:hover {
-  text-decoration: underline; /* Add underline on hover if desired */
-}
-
-.welcome-section {
-  position: relative;
-  background-image: url('../public/first-screen.png');
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-position: center;
-  border: none;
-  outline: none;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  color: #000000;
-}
-
-.welcome-section::after {
-  content: "";
-  position: absolute;
-  bottom: 0; /* Position at the bottom */
-  left: 0;
-  width: 100%;
-  height: 15%; /* Adjust the height of the gradient as desired */
-  background-image: linear-gradient(to bottom, rgba(0, 0, 0, 0), rgba(0, 0, 0, 1));
-}
-
-.welcome-section h1 { /* join us at hackcon */
-  font-size: 64px;
-  margin-top: -250px; /* space above the first line of text */
-  letter-spacing: 5px;
-  line-height: 96px;
-  font-weight: normal;
-}
-
-.welcome-section p { /* embrace ... */
-  color: rgba(0, 0, 0, 0.63);
-  margin-top: -45px;
-  font-size: 24px;
-  letter-spacing: 1px;
-  line-height: 43px;
-  font-weight: normal;
-}
-
-.about-section {
-  background-color: black;
-  text-align: justify;
-}
-
-.content-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.text-container {
-  margin-left: 20px;
-  width: 40%;
-  padding: 40px;
-}
-
-.image-container {
-  width: 50%;
-  margin-right: 150px;
-}
-
-.image-container img {
-  width: 100%;
-  height: auto;
-}
-
-.about-section h1 {
-  color: ghostwhite;
-  font-size: 14px;
-  margin-top: 50px; /* space above the first line of text */
-  letter-spacing: 2px;
-  line-height: 12px;
-  font-weight: normal;
-}
-
-.about-section p {
-  color: ghostwhite;
-  width: 330px;
-  font-size: 14px;
-  letter-spacing: 1px;
-  line-height: 25px;
-  font-weight: normal;
-}
-
-/* vertical line with moving thing */
-.line {
-  position: relative;
-  width: 1px;
-  height: 400px;
-  background-color: rgba(248, 248, 255, 0.8);
-  margin-top: -40px;
-  margin-left: 90px;
-}
-
-.moving-element {
-  position: absolute;
-  top: 0;
-  left: 50%;
-  transform: translateX(-50%) translateY(100px);
-  width: 4px;
-  height: 120px;
-  background-color: white;
-  opacity: 0;
-  animation: move 1.5s 1 linear forwards;
-}
-
-@keyframes move {
-  0%{
-    opacity: 0.4;
-    transform: translateX(-50%) translateY(120px);
-  }
-  100% {
-    opacity: 0.9;
-    transform: translateX(-50%) translateY(40px);
-  }
+    background-color: #f2f2f2;
+    padding: 0;
 }
