+++ src/components/About/about.css
@@ -1,0 +1,78 @@
+.about-section {
+    background-color: black;
+    text-align: justify;
+}
+
+.content-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+}
+
+.text-container {
+    margin-left: 20px;
+    width: 40%;
+    padding: 40px;
+}
+
+.image-container {
+    width: 50%;
+    margin-right: 150px;
+}
+
+.image-container img {
+    width: 100%;
+    height: auto;
+}
+
+.about-section h1 {
+    color: ghostwhite;
+    font-size: 14px;
+    margin-top: 50px; /* space above the first line of text */
+    letter-spacing: 2px;
+    line-height: 12px;
+    font-weight: normal;
+}
+
+.about-section p {
+    color: ghostwhite;
+    width: 330px;
+    font-size: 14px;
+    letter-spacing: 1px;
+    line-height: 25px;
+    font-weight: normal;
+}
+
+/* vertical line with moving thing */
+.line {
+    position: relative;
+    width: 1px;
+    height: 400px;
+    background-color: rgba(248, 248, 255, 0.8);
+    margin-top: -40px;
+    margin-left: 90px;
+}
+
+.moving-element {
+    position: absolute;
+    top: 0;
+    left: 50%;
+    transform: translateX(-50%) translateY(100px);
+    width: 4px;
+    height: 120px;
+    background-color: white;
+    opacity: 0;
+    animation: move 1.5s 1 linear forwards;
+}
+
+@keyframes move {
+    0% {
+        opacity: 0.4;
+        transform: translateX(-50%) translateY(120px);
+    }
+    100% {
+        opacity: 0.9;
+        transform: translateX(-50%) translateY(40px);
+    }
+}
