--- src/components/Navbar/Welcome.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Welcome = () => {
-  return (
-    <div>Welcome</div>
-  )
-}
-
-export default Welcome
