+++ src/components/Navbar/navbar.css
@@ -1,0 +1,47 @@
+.navbar {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    position: fixed;
+    font-size: 16px;
+    letter-spacing: 0;
+    line-height: 50px;
+    top: 0;
+    right: 0;
+    left: 0;
+    padding: 5px 100px;
+    z-index: 1;
+    background-color: transparent;
+    border: none;
+    outline: none;
+}
+
+.navbar.scrolled {
+    background-image: linear-gradient(to bottom, rgba(242, 242, 242, 0.9), rgba(0, 0, 0, 0));
+}
+
+.logo { /* hackcon */
+    font-size: 20px;
+    letter-spacing: 7px;
+    line-height: 17px;
+    font-weight: bold;
+}
+
+.nav-links {
+    cursor: pointer;
+    list-style: none;
+    display: flex;
+}
+
+.nav-links li {
+    margin-left: 35px;
+}
+
+.nav-links li a {
+    text-decoration: none;
+    color: #000000; /* Set the desired color for the links */
+}
+
+.nav-links li a:hover {
+    text-decoration: underline; /* Add underline on hover if desired */
+}
