+++ src/components/Welcome/Welcome.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import './welcome.css'
+
+const Welcome = () => {
+    return (
+        <section id="welcome-section" className="welcome-section">
+            <h1>JOIN US AT HACKCON<br/>2023</h1>
+            <p>Embrace Cutting-Edge Collaborations and Experience<br/>the Future of AI</p>
+        </section>
+    )
+}
+
+export default Welcome
