+++ src/components/Welcome/welcome.css
@@ -1,0 +1,42 @@
+.welcome-section {
+    position: relative;
+    background-image: url('../../assets/first-screen.png');
+    background-size: cover;
+    background-repeat: no-repeat;
+    background-position: center;
+    border: none;
+    outline: none;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    color: #000000;
+}
+
+.welcome-section::after {
+    content: "";
+    position: absolute;
+    bottom: 0; /* Position at the bottom */
+    left: 0;
+    width: 100%;
+    height: 15%; /* Adjust the height of the gradient as desired */
+    background-image: linear-gradient(to bottom, rgba(0, 0, 0, 0), rgba(0, 0, 0, 1));
+}
+
+.welcome-section h1 { /* join us at hackcon */
+    font-size: 64px;
+    margin-top: -250px; /* space above the first line of text */
+    letter-spacing: 5px;
+    line-height: 96px;
+    font-weight: normal;
+}
+
+.welcome-section p { /* embrace ... */
+    color: rgba(0, 0, 0, 0.63);
+    margin-top: -45px;
+    font-size: 24px;
+    letter-spacing: 1px;
+    line-height: 43px;
+    font-weight: normal;
+}
