+++ src/index.css
@@ -2,4 +2,5 @@ body {
     font-family: 'Montserrat', sans-serif;
     margin: 0; /* to delete any borders */
 }
+
 /* Add any additional global styles below */
